+++ src/VSCodeUI.ts
@@ -1,73 +1,127 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license.
-
-import * as os from "os";
-import { Terminal, window, workspace } from "vscode";
+import { Terminal, window, workspace, OutputChannel } from "vscode";
 
 export namespace VSCodeUI{
     export class Commands{
         static cd(cwd: string): string{
-            switch(getTerminalType()){
-                case TerminalType.GitBash:
-                    return `cd "${ cwd.replace(/\\+$/, "") }"`; // Git Bash: remove trailing '\'
-                case TerminalType.CMD:
-                    return `cd /d "${ cwd }"`; // CMD
+            switch (currentWindowsShell()) {
+                case 'Git Bash':
+                    return `cd "${cwd.replace(/\\+$/, "")}"`; // Git Bash: remove trailing '\'
+                case 'PowerShell':
+                    return `cd "${cwd}"`; // PowerShell
+                case 'Command Prompt':
+                    return `cd /d "${cwd}"`; // CMD
+                case 'WSL Bash':
+                    return `cd "${toWSLPath(cwd)}"`; // WSL
                 default:
-                    return `cd "${ cwd }"`;
+                    return `cd "${cwd}"`; // Unknown, try using common one.
             }
         }
         static clearscreen(): string{
-            switch(getTerminalType()){
-                case TerminalType.CMD:
+            switch(currentWindowsShell()){
+                case 'Command Propmt':
                     return "cls";
                 default:
                     return "clear";
             }
         }
         static command(cmd: string): string {
-            switch(getTerminalType()){
-                case TerminalType.PowerShell:
-                    return `& ${ cmd }`; // '&' is call operator.
+            switch (currentWindowsShell()) {
+                case 'PowerShell':
+                    return `cmd /c ${cmd}`; // PowerShell
                 default:
-                    return cmd;
+                    return cmd; // others, try using common one.
             }
         }
     }
 
-    const terminals: { [id: string]: Terminal } = {};
+    export class CompileRunOutputChannel {
+        private readonly channel: OutputChannel = window.createOutputChannel("C/C++ Compile Run");
 
-    export function runInTerminal(command: string, options?: ITerminalOptions) {
-        const defaultOptions: ITerminalOptions = { addNewLine: true, name: "Compile & Run" };
-        const { addNewLine, name, cwd } = Object.assign(defaultOptions, options);
-        if (terminals[name] === undefined)
-            terminals[name] = window.createTerminal({ name });
-        terminals[name].show();
-        if (cwd)
-            terminals[name].sendText(Commands.cd(cwd), true);
-        terminals[name].sendText(Commands.command(command), addNewLine);
-    }
-
-    enum TerminalType{
-        GitBash, PowerShell, CMD, Unknown
-    }
-
-    export function getTerminalType() : TerminalType{
-        if(os.platform() === "win32"){
-            const windowsShell = workspace.getConfiguration("terminal").get<string>("integrated.shell.windows")
-                .toLowerCase(); // The integrated is different from external.
-            if(windowsShell)
-                if (windowsShell.indexOf("bash.exe") > -1 && windowsShell.indexOf("git") > -1)
-                    return TerminalType.GitBash;
-                else if (windowsShell.indexOf("powershell.exe") > -1)
-                    return TerminalType.PowerShell;
-                else if (windowsShell.indexOf("cmd.exe") > -1)
-                    return TerminalType.CMD;
+        public appendLine(message: any, title?: string): void {
+            if (title) {
+                const simplifiedTime: string = (new Date()).toISOString().replace(/z|t/gi, " ").trim(); // YYYY-MM-DD HH:mm:ss.sss
+                const hightlightingTitle: string = `[${title} ${simplifiedTime}]`;
+                this.channel.appendLine(hightlightingTitle);
+            }
+            this.channel.appendLine(message);
         }
-        return TerminalType.Unknown;
+
+        public append(message: any): void {
+            this.channel.append(message);
+        }
+
+        public show(): void {
+            this.channel.show();
+        }
     }
 
-    export function onDidCloseTerminal(closedTerminal: Terminal): void {
-        delete terminals[closedTerminal.name];
+    export class CompileRunTerminal {
+        private readonly terminals: { [id: string]: Terminal } = {};
+
+        public runInTerminal(command: string, options?: ITerminalOptions): void {
+            const defaultOptions: ITerminalOptions = { addNewLine: true, name: "Compile Run" };
+            const { addNewLine, name, cwd } = Object.assign(defaultOptions, options);
+            if (this.terminals[name] === undefined) {
+                this.terminals[name] = window.createTerminal({ name });
+            }
+            this.terminals[name].show();
+            if (cwd) {
+                this.terminals[name].sendText(Commands.command(cwd), true);
+            }
+            this.terminals[name].sendText(Commands.command(command), addNewLine);
+        }
+
+        public closeAllTerminals(): void {
+            Object.keys(this.terminals).forEach((id: string) => {
+                this.terminals[id].dispose();
+                delete this.terminals[id];
+            });
+        }
+
+        public onDidCloseTerminal(closedTerminal: Terminal): void {
+            try {
+                delete this.terminals[closedTerminal.name];
+            } catch (error) {
+                // ignore it.
+            }
+        }
+    }
+
+    export const compileRunTerminal: CompileRunTerminal = new CompileRunTerminal();
+
+    export function currentWindowsShell(): string {
+        const is32ProcessOn64Windows: boolean = process.env.hasOwnProperty('PROCESSOR_ARCHITEW6432');
+        const system32Path: string = `${process.env.windir}\\${is32ProcessOn64Windows ? 'Sysnative' : 'System32'}`;
+        const expectedLocations: { [shell: string]: string[] } = {
+            'Command Prompt': [`${system32Path}\\cmd.exe`],
+            PowerShell: [`${system32Path}\\WindowsPowerShell\\v1.0\\powershell.exe`],
+            'WSL Bash': [`${system32Path}\\bash.exe`],
+            'Git Bash': [
+                `${process.env.ProgramW6432}\\Git\\bin\\bash.exe`,
+                `${process.env.ProgramW6432}\\Git\\usr\\bin\\bash.exe`,
+                `${process.env.ProgramFiles}\\Git\\bin\\bash.exe`,
+                `${process.env.ProgramFiles}\\Git\\usr\\bin\\bash.exe`,
+                `${process.env.LocalAppData}\\Programs\\Git\\bin\\bash.exe`
+            ]
+        };
+        const currentWindowsShellPath: string = workspace.getConfiguration("terminal").get<string>("integrated.shell.windows");
+        for (const key in expectedLocations) {
+            if (expectedLocations[key].indexOf(currentWindowsShellPath) >= 0) {
+                return key;
+            }
+        }
+        return 'Others';
+    }
+
+    export function toWSLPath(p: string): string {
+        const arr: string[] = p.split(":\\");
+        if (arr.length === 2) {
+            const drive: string = arr[0].toLowerCase();
+            const dir: string = arr[1].replace(/\\/g, "/");
+            return `/mnt/${drive}/${dir}`;
+        } else {
+            return ".";
+        }
     }
 }