+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+## [0.2.5]
+- Allow to compile with flags
+- Allow to run with argumets
+- If compiler is not found you can set the new path on while compiling
+
+## [0.2.4]
+- Allow to change compiler path
+- Allow to save before compile
+
+## [0.2.3]
+- Allow to compile or run without doing both
+- Cleanup code
+
+## [0.2.2]
+- Add project links
+- Add donation information
+
+## [0.2.0]
+- Rework compiling method + log
+- Redirect errors to output
+- Add info/error messages
+- Only run in terminal if succeed
+
 ## [0.1.1]
 - Revert to old code